Resolve item search in lines
